Simplify use main hand and move remaining equipment-related stuff to equipment domain
