+++ src/combat/attributes/mana.rs
@@ -16,10 +16,10 @@ impl Mana {
         }
     }
 
-    // optionally uses mana if it can afford it, otherwise returns false if it cost too much
-    pub fn use_mana(&mut self, cost: f32) -> bool {
-        if (self.current_mana - cost) >= 0.0 {
-            self.current_mana -= cost;
+    /** Optionally uses mana if it can afford it, otherwise returns false if it cost too much */
+    pub fn attempt_use_mana(&mut self, cost: &ManaCost) -> bool {
+        if self.current_mana >= cost.0 {
+            self.current_mana -= cost.0;
             return true;
         }
         return false;
@@ -38,3 +38,10 @@ impl Mana {
  */
 #[derive(Component, Clone)]
 pub struct ManaCost(pub f32);
+
+pub fn regenerate_mana(mut query: Query<&mut Mana>, time: Res<Time>) {
+    let delta_time = time.delta_secs();
+    for mut mana in query.iter_mut() {
+        mana.regenerate(delta_time);
+    }
+}
