+++ src/combat/plugin.rs
@@ -1,11 +1,12 @@
 use bevy::prelude::*;
 
 use crate::{
-    combat::{damage::DamagePlugin, status_effects::plugin::StatusEffectPlugin},
+    combat::{
+        attributes::mana::*, damage::DamagePlugin, projectile::handle_collisions::*,
+        status_effects::plugin::StatusEffectPlugin,
+    },
     labels::sets::InGameSet,
 };
-
-use super::projectile::systems::handle_projectile_collisions;
 
 pub struct CombatPlugin;
 
@@ -14,7 +15,10 @@ impl Plugin for CombatPlugin {
         app.add_plugins((DamagePlugin, StatusEffectPlugin))
             .add_systems(
                 Update,
-                handle_projectile_collisions.in_set(InGameSet::Collision),
+                (
+                    regenerate_mana.in_set(InGameSet::Simulation),
+                    handle_projectile_collisions.in_set(InGameSet::Collision),
+                ),
             );
     }
 }
