+++ src/combat/projectile/mod.rs
@@ -1,2 +1,3 @@
 pub mod components;
-pub mod systems;
+pub mod handle_collisions;
+pub mod spawn;
