--- src/combat/weapon/events.rs
@@ -1,4 +1,0 @@
-use bevy::prelude::*;
-
-#[derive(Event)]
-pub struct WeaponAttackTrigger;
