+++ src/combat/weapon/mod.rs
@@ -1,2 +1,1 @@
-pub mod events;
 pub mod weapon;
