+++ src/items/components.rs
@@ -3,25 +3,6 @@ use bevy::prelude::*;
 #[derive(Component, Clone, Debug, Default)]
 pub struct ItemId(pub u32);
 
-#[derive(Component, Clone, PartialEq, Debug)]
-pub enum EquipmentSlot {
-    Mainhand,
-    Helmet,
-}
-
-#[derive(Component, Clone, Debug)]
-pub struct Equippable {
-    pub use_rate: Timer, // swing a sword, shoot a weapon, etc...
-}
-
-impl Default for Equippable {
-    fn default() -> Self {
-        Self {
-            use_rate: Timer::from_seconds(0.4, TimerMode::Once),
-        }
-    }
-}
-
 #[derive(Component, Clone, Debug)]
 pub struct Consumable;
