--- src/items/equipment/components.rs
@@ -1,1 +1,0 @@
-//TODO: Move all  the equipment components here
