+++ src/items/equipment/mod.rs
@@ -1,5 +1,10 @@
 pub mod equip;
-pub mod equipment;
+pub mod equipment_slots;
 pub mod equipment_transform;
 pub mod plugin;
 pub mod unequip;
+pub mod use_equipped;
+
+pub use equip::Equippable;
+pub use equipment_slots::EquipmentSlot;
+pub use equipment_slots::EquipmentSlots;
