--- src/player/systems/regenerate_mana.rs
@@ -1,13 +1,0 @@
-use bevy::{
-    prelude::{Query, Res},
-    time::Time,
-};
-
-use crate::combat::attributes::Mana;
-
-pub fn regenerate_mana(mut query: Query<&mut Mana>, time: Res<Time>) {
-    let delta_time = time.delta_secs();
-    for mut mana in query.iter_mut() {
-        mana.regenerate(delta_time);
-    }
-}
